requirements.txt
fastapi
uvicorn
transformers
sentence-transformers
Pillow
torch
torchvision
requests
python-dotenv
google-cloud-translate
